% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parse_topic_word_weights}
\alias{rcpp_parse_topic_word_weights}
\title{Rcpp Parse Topic Word Weights File --topic-word-weights}
\usage{
rcpp_parse_topic_word_weights(fpath)
}
\arguments{
\item{fpath}{A string specifying path to the word-topic-counts text file.}
}
\value{
topicterms A matrix with topic terms, terms as colnames.
}
\description{
Parses the file generated by Mallet with the --topic-word-weights flag.
The format is a whitespace delimited file, where each line has: topic (starting at 0) term weight.
Refer to the README for this package for more information about the topic-word-weights file.
Note that the weight seems to be count + beta parameter.
}
