% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parse_word_topic_counts}
\alias{rcpp_parse_word_topic_counts}
\title{Rcpp Parse Topic Counts File --word-topic-counts}
\usage{
rcpp_parse_word_topic_counts(fpath, ntopics)
}
\arguments{
\item{fpath}{A string specifying path to the word-topic-counts text file.}

\item{ntopics}{An integer representing the number of topics.}
}
\value{
topicterms A matrix with topic terms, terms as colnames.
}
\description{
Parses the file generated by Mallet with the --word-topic-counts flag.
The format is a whitespace delimited file, where each line has: termnumber term topic:count topic:count ...
Refer to the README for this package for more information about the word-topic-counts file.
Needs to know the number of topics. Returns the topic terms matrix.
}
